+++ application/config/config.php
@@ -1,6 +1,14 @@
 <?php
 defined('BASEPATH') OR exit('No direct script access allowed');
 
+$config['menu_choices'] = array(
+    'menudata' => array(
+        array('name' => 'Home', 'link' => '/'),
+        array('name' => 'Profile', 'link' => '/profile'),
+        array('name' => 'Stocks', 'link' => '/stock'),
+    )
+);
+
 /*
 |--------------------------------------------------------------------------
 | Base Site URL