+++ application/controllers/Welcome.php
@@ -1,25 +1,37 @@
 <?php
-defined('BASEPATH') OR exit('No direct script access allowed');
 
-class Welcome extends CI_Controller {
+/**
+ * Our homepage. Show a table of all the author pictures. Clicking on one should show their quote.
+ * Our quotes model has been autoloaded, because we use it everywhere.
+ *
+ * controllers/Welcome.php
+ *
+ * ------------------------------------------------------------------------
+ */
+class Welcome extends Application {
 
-	/**
-	 * Index Page for this controller.
-	 *
-	 * Maps to the following URL
-	 * 		http://example.com/index.php/welcome
-	 *	- or -
-	 * 		http://example.com/index.php/welcome/index
-	 *	- or -
-	 * Since this controller is set as the default controller in
-	 * config/routes.php, it's displayed at http://example.com/
-	 *
-	 * So any other public methods not prefixed with an underscore will
-	 * map to /index.php/welcome/<method_name>
-	 * @see https://codeigniter.com/user_guide/general/urls.html
-	 */
-	public function index()
-	{
-		$this->load->view('welcome_message');
-	}
+    function __construct() {
+        parent::__construct();
+
+        $this->load->model('players');
+        $this->load->model('stocks');
+    }
+
+    //-------------------------------------------------------------
+    //  The normal pages
+    //-------------------------------------------------------------
+
+    function index() {
+        $this->data['pageTitle'] = 'Homepage';
+        // this is the view we want shown
+        $this->data['pagebody'] = 'homepage';
+
+        $this->data['players'] = $this->players->all();
+        $this->data['stocks'] = $this->stocks->all();
+
+        $this->render();
+    }
 }
+
+/* End of file Welcome.php */
+/* Location: application/controllers/Welcome.php */