+++ application/views/_menubar.php
@@ -1,0 +1,8 @@
+<?php
+/*
+ * Menu navbar, just an unordered list
+ */
+?>
+{menudata}
+<li><a href="{link}">{name}</a></li>
+{/menudata}